Move translation loading into TranslationProvider
