+++ src/Translation/components/TranslationContext.tsx
@@ -1,7 +1,9 @@
-import React, { createContext, useState } from "react";
+import React, { createContext, useEffect, useState } from "react";
 
-import { DefaultLocale, Locale } from "@translation/model";
+import { flattenObject } from '@core/helpers';
 import { Dictionary } from "@shared/types";
+import { DefaultLocale, Locale } from "@translation/model";
+import appTranslations from "@translation/index";
 
 type Context = {
 	locale: Locale,
@@ -25,8 +27,42 @@ const TranslationProvider: React.FunctionComponent<Props> = (props: Props) => {
 	const [locale, setLocale] = useState(DefaultLocale);
 	const [translations, setTranslations] = useState({});
 
+	useEffect(() => {
+		if (Object.keys(translations).length === 0) {
+			loadTranslations(locale);
+		}
+	}, [translations]);
+
+	const getLocaleTranslations = (locale: string): object => appTranslations[locale as keyof typeof appTranslations];
+
+	const hasLocale = (locale: string): boolean => getLocaleTranslations(locale) !== null;
+
+	const loadTranslations = (locale: string): void => {
+		const translationResources = getLocaleTranslations(locale);
+
+		if (!translationResources) {
+			throw new Error(`Failed to load translation resources for locale '${locale}'.`);
+		}
+
+		setTranslations(flattenObject(translationResources) as Dictionary<string>);
+	};
+
+	const handleSetLocale = (newLocale: Locale): void => {
+		if (hasLocale(newLocale)) {
+			setLocale(newLocale);
+			loadTranslations(newLocale);
+		}
+		else {
+			console.error(`Locale "${newLocale}" do not exist`);
+		}
+	};
+
 	return (
-		<TranslationContext.Provider value={{ locale, setLocale, translations, setTranslations } as Context}>
+		<TranslationContext.Provider value={{
+			locale,
+			setLocale: handleSetLocale,
+			translations
+		} as Context}>
 			{props.children}
 		</TranslationContext.Provider>
 	);
