+++ src/Translation/hooks/useLocale.ts
@@ -1,10 +1,7 @@
-import { Dictionary } from '@shared/types';
 import { useContext } from "react";
 
 import { Locale } from "@translation/model";
 import { TranslationContext } from "@translation/components/TranslationContext";
-import appTranslations from "@translation/index";
-import { flattenObject } from '@core/helpers';
 
 type UseLocale = {
 	locale: Locale,
@@ -13,7 +10,7 @@ type UseLocale = {
 };
 
 const useLocale = (): UseLocale => {
-	const { locale, setLocale, translations, setTranslations } = useContext(TranslationContext);
+	const { locale, setLocale, translations } = useContext(TranslationContext);
 
 	const translate = (key: string): string => {
 		const translatedText = translations[key];
@@ -27,31 +24,7 @@ const useLocale = (): UseLocale => {
 		return translatedText;
 	};
 
-	const handleSetLocale = (newLocale: Locale): void => {
-		if (hasLocale(newLocale)) {
-			setLocale(newLocale);
-			loadTranslations(newLocale);
-		}
-		else {
-			console.error(`Locale "${newLocale}" do not exist`);
-		}
-	};
-
-	const loadTranslations = (locale: string): void => {
-		const translationResources = getLocaleTranslations(locale);
-
-		if (!translationResources) {
-			throw new Error(`Failed to load translation resources for locale '${locale}'.`);
-		}
-
-		setTranslations(flattenObject(translationResources) as Dictionary<string>);
-	};
-
-	const getLocaleTranslations = (locale: string): object => appTranslations[locale as keyof typeof appTranslations];
-
-	const hasLocale = (locale: string): boolean => getLocaleTranslations(locale) !== null;
-
-	return { locale, setLocale: handleSetLocale, translate };
+	return { locale, setLocale, translate };
 };
 
 export default useLocale;
